+++ src/types/esm.ts
@@ -1,3 +1,3 @@
 export interface EsmService {
-  load(url: string, depMap: Record<string, string>): Promise<any>;
+  load(url: string, depMap: Record<string, string>, parent: string): Promise<any>;
 }